Separate mockRDSClient code to another file
